% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{Title}
\usage{
activity(expression_matrix, adjacency_edges = .data$differentiation_edges)
}
\arguments{
\item{expression_matrix}{Input expression matrix with cells as columns and genes as rows(matrix or data frame)}

\item{adjacency_edges}{Adjacency edges of the protein protein interaction network}
}
\value{
A vector with the differentiation activity values for each cell
}
\description{
Title
}
