% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_edges.R
\docType{data}
\name{differentiation_edges}
\alias{differentiation_edges}
\title{Adjacency edges.}
\format{
A data frame with 191072 rows and 2 variables:
\describe{
\item{V1}{node one}
\item{V2}{node two}
...
}
}
\source{
\url{https://www.ebi.ac.uk/QuickGO/ and https://www.pathwaycommons.org/}
}
\usage{
differentiation_edges
}
\description{
A dataset containing the edges of the differentiation protein-protein interaction network
}
\keyword{datasets}
